requirements.txt
streamlit
pandas
matplotlib
Pillow
sqlalchemy
BytesIO
Image
ImageDraw
ImageFont